refactor(tools): re-position test files
