+++ packages/tools/src/findDupeFileFromGlob.test.ts
@@ -4,10 +4,10 @@ import { expect, jest, test } from "@jest/globals";
 import memfs, { vol } from "memfs";
 import process from "process";
 
-import { findDupeFileSet } from "../findDupeFileSet";
-import { findFilePathsFromGlob } from "../findFilePathsFromGlob";
-import { readAssetBoxConfig } from "../readAssetBoxConfig";
-import { createMockRepository } from "./utils/mockRepository";
+import { findDupeFileSet } from "./findDupeFileSet";
+import { findFilePathsFromGlob } from "./findFilePathsFromGlob";
+import { readAssetBoxConfig } from "./readAssetBoxConfig";
+import { createMockRepository } from "./testUtils/mockRepository";
 
 const spy = jest.spyOn(process, "cwd");
